requirements.txt
PyMuPDF
langchain
langchain-community
cohere
python-dotenv
chromadb
fastapi
uvicorn